+++ d3blocks/violin/Violin.py
@@ -6,7 +6,7 @@ import pandas as pd
 from jinja2 import Environment, PackageLoader
 from pathlib import Path
 import os
-
+import time
 
 def preprocessing(x, y, config, c=None, s=5, stroke='#ffffff', opacity=0.8, tooltip='', logger=None):
     # Convert to dataframe
@@ -82,7 +82,7 @@ def show(df, config, labels=None):
     return config
 
 
-def write_html(X, config, overwrite=True):
+def write_html(X, config):
     """Write html.
 
     Parameters
@@ -116,12 +116,11 @@ def write_html(X, config, overwrite=True):
     jinja_env = Environment(loader=PackageLoader(package_name=__name__, package_path='d3js'))
     index_template = jinja_env.get_template('violin.html.j2')
     index_file = Path(config['filepath'])
-    print('Write to path: [%s]' % index_file.absolute())
     # index_file.write_text(index_template.render(content))
-    if os.path.isfile(index_file):
-        if overwrite:
-            print('File already exists and will be overwritten: [%s]' %(index_file))
-            os.remove(index_file)
+    if config['overwrite'] and os.path.isfile(index_file):
+        print('File already exists and will be overwritten: [%s]' %(index_file))
+        os.remove(index_file)
+        time.sleep(0.5)
     with open(index_file, "w", encoding="utf-8") as f:
         f.write(index_template.render(content))
